rename py files to entity_recognition.
